Move all files related to the IDE to `ide`
