+++ src/ide/codemirror-tscriptmode.ts
