+++ src/ide/search.ts
