+++ src/ide/standalone.ts
